+++ services/common/ir/parsers.py
@@ -243,11 +243,9 @@ class ParserInforetrieval(Parser):
         try:
             if json_input.get("generic"):
                 # Retrocompatibility with older infoindexing calls
-                self.index_conf = get_index_conf(generic=json_input.get("generic"))
+                self.index_conf = json_input["generic"].get("index_conf")
             else:
-                self.index_conf = json_input.get("index_conf")
-                if not self.index_conf:
-                    self.index_conf = json_input.get("indexation  _conf")
+                self.index_conf = json_input["indexation_conf"]
 
             self.get_index(self.index_conf)
             self.get_filters(self.index_conf)